Finish array cardio day 2!
Learned how the some method will return true if it finds one
truthy value in the array and return false is none are found.
The every method withh return false if it finds one falsy value and
return true is all items in array are true.

Also got good practice with find and findIndex methods!